+++ app/build.gradle
@@ -12,8 +12,8 @@ android {
         applicationId "com.sistemas51.horarioslavalle"
         minSdkVersion 19
         targetSdkVersion 29
-        versionCode 32
-        versionName "2.5.3"
+        versionCode 34
+        versionName "2.6.3"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables.useSupportLibrary = true
         dataBinding.enabled false
@@ -32,7 +32,7 @@ dependencies {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
     implementation "androidx.appcompat:appcompat:1.1.0"
-    def nav_version = "2.1.0-rc01"
+    def nav_version = "2.2.0"
     implementation "androidx.navigation:navigation-fragment:$nav_version"
     implementation "androidx.navigation:navigation-ui:$nav_version"
