+++ app/src/main/java/com/sistemas51/horarioslavalle/fragments/DestinySelected.java
@@ -1,22 +1,20 @@
-package com.sistemas51.horarioslavalle.router;
+package com.sistemas51.horarioslavalle.fragments;
 
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 
-import androidx.activity.OnBackPressedCallback;
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.sistemas51.horarioslavalle.R;
-import com.sistemas51.horarioslavalle.UtilidadesAdaptadores.HorarioModel;
-import com.sistemas51.horarioslavalle.UtilidadesAdaptadores.StepperRvAdapter;
+import com.sistemas51.horarioslavalle.models.HorarioModel;
+import com.sistemas51.horarioslavalle.adapters.StepperRvAdapter;
 
 import java.util.List;
 import java.util.Map;
@@ -24,13 +22,11 @@ import java.util.Map;
 /**
  * A simple {@link Fragment} subclass.
  */
-public class OriginSelected extends Fragment {
-
-
-    private RecyclerView rv;
-    private StepperRvAdapter stepperRvAdapter;
-    private int arrayId;
-    public OriginSelected() {
+public class DestinySelected extends Fragment {
+    RecyclerView rv;
+    StepperRvAdapter stepperRvAdapter;
+    Toolbar toolbar;
+    public DestinySelected() {
         // Required empty public constructor
     }
 
@@ -39,17 +35,19 @@ public class OriginSelected extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         View v = inflater.inflate(R.layout.fragment_listview, container, false);
-        ((AppCompatActivity) getActivity()).getSupportActionBar().setTitle("Seleccionar Origen");
+        rv = (RecyclerView) v.findViewById(R.id.recicler);
+        ((AppCompatActivity) getActivity()).getSupportActionBar().setTitle("Seleccionar Destino");
         ((AppCompatActivity) getActivity()).getSupportActionBar().setSubtitle("Horarios Lavalle");
 
-
-        rv = (RecyclerView) v.findViewById(R.id.recicler);
-        Map<String, String> args = OriginSelectedArgs.fromBundle(getArguments()).getArgs();
+        Map<String,String> args = OriginSelectedArgs.fromBundle(getArguments()).getArgs();
         String routeSelected = args.get(getString(R.string.route));
         List<String> places = HorarioModel.getPlacesFromRoute(routeSelected,getContext());
         rv.setLayoutManager(new LinearLayoutManager(getContext(), LinearLayoutManager.VERTICAL, false));
-        stepperRvAdapter = new StepperRvAdapter(getContext(),places,1, args);
+        stepperRvAdapter = new StepperRvAdapter(getContext(),places,2, args);
+        stepperRvAdapter.notifyDataSetChanged();
         rv.setAdapter(stepperRvAdapter);
+
         return v;
     }
+
 }
