--- app/src/main/java/com/sistemas51/horarioslavalle/models/WeatherModel.java
@@ -1,43 +1,0 @@
-package com.sistemas51.horarioslavalle.models;
-
-public class WeatherModel {
-
-    String place;
-    String description;
-    String current;
-
-    String image;
-
-    public String getPlace() {
-        return place;
-    }
-
-    public void setPlace(String place) {
-        this.place = place;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getCurrent() {
-        return current;
-    }
-
-    public void setCurrent(String current) {
-        this.current = current;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-}
-
