+++ app/src/main/res/layout/activity_ayuda.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.sistemas51.horarioslavalle.UtilidadesAdaptadores.Help">
+    tools:context="com.sistemas51.horarioslavalle.activity.Help">
     <androidx.appcompat.widget.Toolbar
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
