+++ app/src/main/res/layout/activity_special_hours.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".v2.SpecialHours">
+    tools:context=".activity.SpecialHours">
 
     <WebView
         android:layout_width="match_parent"
