+++ app/src/main/res/layout/fragment_listview.xml
@@ -4,22 +4,12 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_height="match_parent">
 
-    <!-- TODO: Update blank fragment layout -->
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recicler"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
 />
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/floating_action_button"
-        android:visibility="gone"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="end|bottom"
-        android:layout_margin="16dp"
-        app:srcCompat="@drawable/ic_wb_sunny_black_24dp"/>
 
-    <!-- Adding bottom sheet after main content -->
-    <include layout="@layout/weather_reclycler_view" />
+
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
