--- app/src/main/res/layout/weather_reclycler_view.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:id="@+id/bottomSheet"
-    android:layout_height="wrap_content"
-    app:layout_behavior="@string/bottom_sheet_behavior">
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/weatherRecycler"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        />
-</LinearLayout>
