--- app/src/main/res/layout/weather_view.xml
@@ -1,79 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <androidx.cardview.widget.CardView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:layout_marginLeft="10dp"
-        android:layout_marginRight="10dp"
-        android:layout_marginBottom="10dp">
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical">
-
-            <ImageView
-                android:id="@+id/imageTemperature"
-                android:layout_width="60dp"
-                android:layout_height="60dp"
-                android:layout_centerVertical="true"
-                android:layout_marginTop="10dp"
-                android:layout_marginRight="10dp"
-                android:layout_alignParentRight="true"
-                tools:src="@drawable/help" />
-
-
-            <TextView
-                android:id="@+id/place"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginLeft="20dp"
-                android:textSize="18dp"
-                android:layout_alignParentLeft="true"
-                android:layout_marginTop="10dp"
-                android:layout_toLeftOf="@id/imageTemperature"
-                android:textColor="@color/black"
-                tools:text="Gustavo Andre"
-                />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:id="@+id/description"
-                android:layout_below="@id/place"
-                android:layout_marginLeft="20dp"
-                android:layout_marginTop="10dp"
-                tools:text="Soleado"/>
-
-            <TextView
-                android:id="@+id/text_current"
-                android:visibility="gone"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_below="@id/description"
-                android:layout_marginLeft="20dp"
-                android:text="Actual"
-                android:textColor="@color/colorAccent"
-                android:layout_marginTop="10dp"
-                />
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:id="@+id/currentTemperature"
-                android:layout_below="@id/description"
-                android:layout_marginLeft="20dp"
-                android:textSize="40dp"
-                android:textColor="@color/colorAccent"
-                tools:text="28°C"
-                android:layout_marginTop="20dp"
-                />
-
-
-        </RelativeLayout>
-    </androidx.cardview.widget.CardView>
-
-</LinearLayout>
